+++ packages/shield-controller/CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Bump `@metamask/transaction-controller` from `^60.7.0` to `^60.8.0` ([#6883](https://github.com/MetaMask/core/pull/6883))
+- Updated internal coverage result polling and log logic. ([#6847](https://github.com/MetaMask/core/pull/6847))
+  - Added cancellation logic to the polling.
+  - Updated implementation of timeout.
+  - Cancel any pending requests before starting new polling or logging.
+- Updated TransactionMeta comparison in `TransactionController:stateChange` subscriber to avoid triggering multiple check coverage result unnecessarily. ([#6847](https://github.com/MetaMask/core/pull/6847))
+- Removed `Personal Sign` check from the check signature coverage result. ([#6847](https://github.com/MetaMask/core/pull/6847))
 
 ## [0.3.2]